Two examples, WildFunction and BesselRoots, for the group demo.
WildFunction highlights some serious flaws in fun_roots_unit.
For one, we usume the length of the restricted funs will always
be less than half of the orignial. This is rarely the case.
In some cases the length doesn't even decrease at all, 
suggesting that a horizontal scale isn't being taken into
account somewhere.

As a work-around, the magic constant 'c' from roots has been
dropped, and we force restricted funs to be at most half the 
length of their parents. Since PG plans to implement midpoint
splitting and a recurrence relation for the child funs, this
should be fine as a temporary fix.